+++ src/textual/pilot.py
@@ -14,7 +14,12 @@ from .widget import Widget
 
 
 def _get_mouse_message_arguments(
-    target: Widget, offset: Offset = Offset(), button: int = 0
+    target: Widget,
+    offset: Offset = Offset(),
+    button: int = 0,
+    shift: bool = False,
+    meta: bool = False,
+    control: bool = False,
 ) -> dict[str, Any]:
     """Get the arguments to pass into mouse messages for the click and hover methods."""
     click_x, click_y, _, _ = target.region.translate(offset)
@@ -24,9 +29,9 @@ def _get_mouse_message_arguments(
         "delta_x": 0,
         "delta_y": 0,
         "button": button,
-        "shift": False,
-        "meta": False,
-        "ctrl": False,
+        "shift": shift,
+        "meta": meta,
+        "ctrl": control,
         "screen_x": click_x,
         "screen_y": click_y,
     }
@@ -59,7 +64,12 @@ class Pilot(Generic[ReturnType]):
             await self._app._press_keys(keys)
 
     async def click(
-        self, selector: QueryType | None = None, offset: Offset = Offset()
+        self,
+        selector: QueryType | None = None,
+        offset: Offset = Offset(),
+        shift: bool = False,
+        meta: bool = False,
+        control: bool = False,
     ) -> None:
         """Simulate clicking with the mouse.
 
@@ -70,6 +80,9 @@ class Pilot(Generic[ReturnType]):
                 currently hidden or obscured by another widget, then the click may
                 not land on it.
             offset: The offset to click within the selected widget.
+            shift: Click with the shift key held down.
+            meta: Click with the meta key held down.
+            control: Click with the control key held down.
         """
         app = self.app
         screen = app.screen
@@ -79,7 +92,7 @@ class Pilot(Generic[ReturnType]):
             target_widget = screen
 
         message_arguments = _get_mouse_message_arguments(
-            target_widget, offset, button=1
+            target_widget, offset, button=1, shift=shift, meta=meta, control=control
         )
         app.post_message(MouseDown(**message_arguments))
         app.post_message(MouseUp(**message_arguments))